+++ src/ransac_impl/ransac_impl.h
@@ -23,13 +23,13 @@ public:
         threshold = thresh;
     };
     void setMaxIteration(int in){maxIteration = in;};
-    void setThreshold(int in){threshold = in;};
+    void setThreshold(double in){threshold = in;};
     void run();
     std::unordered_set<int> RansacLine(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud, 
 								 double maxIterations, double distanceTol);
     
     std::unordered_set<int> RansacPlane(pcl::PointCloud<pcl::PointXYZ>::Ptr cloud);
-    
+
 
 
 private:
